server/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
pydantic>=2.7.4
python-multipart>=0.0.9
scikit-learn>=1.3.2
xgboost>=2.0.3
joblib>=1.4.2
PyPDF2>=3.0.1
python-docx>=1.1.2
sentence-transformers>=3.0.0
